@OneToOne(optinal = true) @Test 수정 (영속성 고려 중요) && Entity Listener 추가

+++ src/test/java/com/fintech/chatbot_spring/Dao/MessageRepositoryTest.java
@@ -1,13 +1,10 @@
 package com.fintech.chatbot_spring.Dao;
 
-import org.aspectj.lang.annotation.After;
 import org.junit.jupiter.api.*;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
 import com.fintech.chatbot_spring.Domain.Message;
 import java.util.List;
 
@@ -30,10 +27,16 @@ public class MessageRepositoryTest {
     @Test
     @Order(1)
     public void testSaveThenFind() {
-        Long id = (long)1;
+
+        Long id = 1L;
         String content = "첫 번째 메세지";
 
-        messageRepository.save(Message.builder().id(id).content(content).build());
+        messageRepository.save(Message
+                .builder()
+                .id(id)
+                .content(content)
+                .build()
+        );
 
         List<Message> messageList = messageRepository.findAll();
         logger.info(messageList.toString());
